Update index.qmd
I think the included changes should work w v.4.0.0 of ggplot2. I also decided that using na.rm = T argument was more true to the specification of some reasonable compute, so other changes relate to that.   Let me know if I've introduced any problems, and I'll be happy to try to address.  Thank you so much for this highlight.  It's great!  🤩🧡